backend/requirements.txt
aiofiles==23.2.1
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
arxiv==2.3.1
async-timeout==4.0.3
audio-recorder-streamlit==0.0.10
audioread==3.1.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==5.5.2
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
ddgs==9.9.1
decorator==5.2.1
distro==1.9.0
entrypoints==0.4
fake-useragent==2.2.0
fastapi==0.115.2
feedparser==6.0.12
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
gradio==4.36.1
gradio_client==1.0.1
greenlet==3.2.4
hf-xet==1.2.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.23.4
importlib-metadata==6.11.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.0.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.7
langchain-experimental==0.4.0
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.44
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.45.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
msgpack==1.1.2
mypy_extensions==1.1.0
narwhals==2.12.0
numba==0.62.1
numpy==1.26.4
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==23.2
pandas==2.3.3
pillow==10.4.0
platformdirs==4.5.0
pooch==1.8.2
primp==0.15.0
protobuf==4.25.8
pyarrow==22.0.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.3
redis==5.0.7
referencing==0.37.0
regex==2025.11.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.29.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
ruff==0.14.5
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
socksio==1.0.0
soundfile==0.12.1
soupsieve==2.8
soxr==1.0.0
SQLAlchemy==2.0.44
starlette==0.40.0
streamlit==1.28.1
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
tomlkit==0.12.0
toolz==1.1.0
tornado==6.5.2
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.30.6
uvloop==0.22.1
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==12.0
wikipedia==1.4.0
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
